+++ src/commands/concat_commands.rs
@@ -28,36 +28,33 @@ fn concat_into_string(
     interpreter: &mut Interpreter,
     mut command: Command,
     conversion_fn: impl Fn(&str) -> String,
-) -> Result<TokenStream> {
-    let interpreted =
-        command.interpret_remaining_arguments(interpreter, SubstitutionMode::token_stream())?;
+) -> Result<InterpretedStream> {
+    let interpreted = command.arguments().interpret_as_tokens(interpreter)?;
     let concatenated = concat_recursive(interpreted);
     let string_literal = string_literal(&conversion_fn(&concatenated), command.span());
-    Ok(TokenStream::from(TokenTree::Literal(string_literal)))
+    Ok(InterpretedStream::of_literal(string_literal))
 }
 
 fn concat_into_ident(
     interpreter: &mut Interpreter,
     mut command: Command,
     conversion_fn: impl Fn(&str) -> String,
-) -> Result<TokenStream> {
-    let interpreted =
-        command.interpret_remaining_arguments(interpreter, SubstitutionMode::token_stream())?;
+) -> Result<InterpretedStream> {
+    let interpreted = command.arguments().interpret_as_tokens(interpreter)?;
     let concatenated = concat_recursive(interpreted);
     let ident = parse_ident(&conversion_fn(&concatenated), command.span())?;
-    Ok(TokenStream::from(TokenTree::Ident(ident)))
+    Ok(InterpretedStream::of_ident(ident))
 }
 
 fn concat_into_literal(
     interpreter: &mut Interpreter,
     mut command: Command,
     conversion_fn: impl Fn(&str) -> String,
-) -> Result<TokenStream> {
-    let interpreted =
-        command.interpret_remaining_arguments(interpreter, SubstitutionMode::token_stream())?;
+) -> Result<InterpretedStream> {
+    let interpreted = command.arguments().interpret_as_tokens(interpreter)?;
     let concatenated = concat_recursive(interpreted);
     let literal = parse_literal(&conversion_fn(&concatenated), command.span())?;
-    Ok(TokenStream::from(TokenTree::Literal(literal)))
+    Ok(InterpretedStream::of_literal(literal))
 }
 
 //=======================================
@@ -69,7 +66,7 @@ pub(crate) struct StringCommand;
 impl CommandDefinition for StringCommand {
     const COMMAND_NAME: &'static str = "string";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, |s| s.to_string())
     }
 }
@@ -79,7 +76,7 @@ pub(crate) struct IdentCommand;
 impl CommandDefinition for IdentCommand {
     const COMMAND_NAME: &'static str = "ident";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_ident(interpreter, command, |s| s.to_string())
     }
 }
@@ -89,7 +86,7 @@ pub(crate) struct IdentCamelCommand;
 impl CommandDefinition for IdentCamelCommand {
     const COMMAND_NAME: &'static str = "ident_camel";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_ident(interpreter, command, to_upper_camel_case)
     }
 }
@@ -99,7 +96,7 @@ pub(crate) struct IdentSnakeCommand;
 impl CommandDefinition for IdentSnakeCommand {
     const COMMAND_NAME: &'static str = "ident_snake";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_ident(interpreter, command, to_lower_snake_case)
     }
 }
@@ -109,7 +106,7 @@ pub(crate) struct IdentUpperSnakeCommand;
 impl CommandDefinition for IdentUpperSnakeCommand {
     const COMMAND_NAME: &'static str = "ident_upper_snake";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_ident(interpreter, command, to_upper_snake_case)
     }
 }
@@ -119,7 +116,7 @@ pub(crate) struct LiteralCommand;
 impl CommandDefinition for LiteralCommand {
     const COMMAND_NAME: &'static str = "literal";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_literal(interpreter, command, |s| s.to_string())
     }
 }
@@ -133,7 +130,7 @@ pub(crate) struct UpperCommand;
 impl CommandDefinition for UpperCommand {
     const COMMAND_NAME: &'static str = "upper";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_uppercase)
     }
 }
@@ -143,7 +140,7 @@ pub(crate) struct LowerCommand;
 impl CommandDefinition for LowerCommand {
     const COMMAND_NAME: &'static str = "lower";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_lowercase)
     }
 }
@@ -153,7 +150,7 @@ pub(crate) struct SnakeCommand;
 impl CommandDefinition for SnakeCommand {
     const COMMAND_NAME: &'static str = "snake";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         // Lower snake case is the more common casing in Rust, so default to that
         LowerSnakeCommand::execute(interpreter, command)
     }
@@ -164,7 +161,7 @@ pub(crate) struct LowerSnakeCommand;
 impl CommandDefinition for LowerSnakeCommand {
     const COMMAND_NAME: &'static str = "lower_snake";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_lower_snake_case)
     }
 }
@@ -174,7 +171,7 @@ pub(crate) struct UpperSnakeCommand;
 impl CommandDefinition for UpperSnakeCommand {
     const COMMAND_NAME: &'static str = "upper_snake";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_upper_snake_case)
     }
 }
@@ -184,7 +181,7 @@ pub(crate) struct KebabCommand;
 impl CommandDefinition for KebabCommand {
     const COMMAND_NAME: &'static str = "kebab";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         // Kebab case is normally lower case (including in Rust where it's used - e.g. crate names)
         // It can always be combined with other casing to get other versions
         concat_into_string(interpreter, command, to_lower_kebab_case)
@@ -196,7 +193,7 @@ pub(crate) struct CamelCommand;
 impl CommandDefinition for CamelCommand {
     const COMMAND_NAME: &'static str = "camel";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         // Upper camel case is the more common casing in Rust, so default to that
         UpperCamelCommand::execute(interpreter, command)
     }
@@ -207,7 +204,7 @@ pub(crate) struct LowerCamelCommand;
 impl CommandDefinition for LowerCamelCommand {
     const COMMAND_NAME: &'static str = "lower_camel";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_lower_camel_case)
     }
 }
@@ -217,7 +214,7 @@ pub(crate) struct UpperCamelCommand;
 impl CommandDefinition for UpperCamelCommand {
     const COMMAND_NAME: &'static str = "upper_camel";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, to_upper_camel_case)
     }
 }
@@ -227,7 +224,7 @@ pub(crate) struct CapitalizeCommand;
 impl CommandDefinition for CapitalizeCommand {
     const COMMAND_NAME: &'static str = "capitalize";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, capitalize)
     }
 }
@@ -237,7 +234,7 @@ pub(crate) struct DecapitalizeCommand;
 impl CommandDefinition for DecapitalizeCommand {
     const COMMAND_NAME: &'static str = "decapitalize";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, decapitalize)
     }
 }
@@ -247,7 +244,7 @@ pub(crate) struct TitleCommand;
 impl CommandDefinition for TitleCommand {
     const COMMAND_NAME: &'static str = "title";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, title_case)
     }
 }
@@ -257,12 +254,12 @@ pub(crate) struct InsertSpacesCommand;
 impl CommandDefinition for InsertSpacesCommand {
     const COMMAND_NAME: &'static str = "insert_spaces";
 
-    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<TokenStream> {
+    fn execute(interpreter: &mut Interpreter, command: Command) -> Result<InterpretedStream> {
         concat_into_string(interpreter, command, insert_spaces_between_words)
     }
 }
 
-fn concat_recursive(arguments: TokenStream) -> String {
+fn concat_recursive(arguments: InterpretedStream) -> String {
     fn concat_recursive_internal(output: &mut String, arguments: TokenStream) {
         for token_tree in arguments {
             match token_tree {
@@ -307,6 +304,6 @@ fn concat_recursive(arguments: TokenStream) -> String {
     }
 
     let mut output = String::new();
-    concat_recursive_internal(&mut output, arguments);
+    concat_recursive_internal(&mut output, arguments.into_token_stream());
     output
 }